+++ frontend/src/pages/AdDetail.tsx
